+++ src/components/radio/radio.ts
@@ -1,11 +1,9 @@
-import { html, LitElement } from 'lit';
-import { customElement, property, query, state } from 'lit/decorators.js';
+import { html } from 'lit';
+import { customElement } from 'lit/decorators.js';
 import { classMap } from 'lit/directives/class-map.js';
 import { ifDefined } from 'lit/directives/if-defined.js';
 import { live } from 'lit/directives/live.js';
-import { emit } from '~/internal/event';
-import { FormSubmitController } from '~/internal/form-control';
-import { watch } from '~/internal/watch';
+import RadioBase from '~/internal/radio';
 import styles from './radio.styles';
 
 /**
@@ -24,111 +22,8 @@ import styles from './radio.styles';
  * @csspart label - The radio label.
  */
 @customElement('sl-radio')
-export default class SlRadio extends LitElement {
+export default class SlRadio extends RadioBase {
   static styles = styles;
-
-  @query('input[type="radio"]') input: HTMLInputElement;
-
-  // @ts-expect-error -- Controller is currently unused
-  private readonly formSubmitController = new FormSubmitController(this, {
-    value: (control: SlRadio) => (control.checked ? control.value : undefined)
-  });
-
-  @state() private hasFocus = false;
-
-  /** The radio's name attribute. */
-  @property() name: string;
-
-  /** The radio's value attribute. */
-  @property() value: string;
-
-  /** Disables the radio. */
-  @property({ type: Boolean, reflect: true }) disabled = false;
-
-  /** Draws the radio in a checked state. */
-  @property({ type: Boolean, reflect: true }) checked = false;
-
-  /**
-   * This will be true when the control is in an invalid state. Validity in range inputs is determined by the message
-   * provided by the `setCustomValidity` method.
-   */
-  @property({ type: Boolean, reflect: true }) invalid = false;
-
-  connectedCallback(): void {
-    super.connectedCallback();
-    this.setAttribute('role', 'radio');
-  }
-
-  /** Simulates a click on the radio. */
-  click() {
-    this.input.click();
-  }
-
-  /** Sets focus on the radio. */
-  focus(options?: FocusOptions) {
-    this.input.focus(options);
-  }
-
-  /** Removes focus from the radio. */
-  blur() {
-    this.input.blur();
-  }
-
-  /** Checks for validity and shows the browser's validation message if the control is invalid. */
-  reportValidity() {
-    return this.input.reportValidity();
-  }
-
-  /** Sets a custom validation message. If `message` is not empty, the field will be considered invalid. */
-  setCustomValidity(message: string) {
-    this.input.setCustomValidity(message);
-    this.invalid = !this.input.checkValidity();
-  }
-
-  getAllRadios() {
-    const radioGroup = this.closest('sl-radio-group');
-
-    // Radios must be part of a radio group
-    if (radioGroup === null) {
-      return [this];
-    }
-
-    return [...radioGroup.querySelectorAll<SlRadio>('sl-radio')].filter((radio: this) => radio.name === this.name);
-  }
-
-  handleBlur() {
-    this.hasFocus = false;
-    emit(this, 'sl-blur');
-  }
-
-  @watch('checked')
-  handleCheckedChange() {
-    this.setAttribute('aria-checked', this.checked ? 'true' : 'false');
-
-    if (this.hasUpdated) {
-      emit(this, 'sl-change');
-    }
-  }
-
-  handleClick() {
-    this.checked = true;
-  }
-
-  @watch('disabled', { waitUntilFirstUpdate: true })
-  handleDisabledChange() {
-    this.setAttribute('aria-disabled', this.disabled ? 'true' : 'false');
-
-    // Disabled form controls are always valid, so we need to recheck validity when the state changes
-    if (this.hasUpdated) {
-      this.input.disabled = this.disabled;
-      this.invalid = !this.input.checkValidity();
-    }
-  }
-
-  handleFocus() {
-    this.hasFocus = true;
-    emit(this, 'sl-focus');
-  }
 
   render() {
     return html`